MercSheet semi-integrated to save system
Using MercSheets and some logic in PlayerData, adding mercs is a proper method now, which takes MercName(enums) and CreateAdds them to the save file.
It can also just Add an existing sheet (for hire-able mercs) 

Sheets have methods of returning the prefab to their Pawn object, and the can also return whichever component is on that prefab's Root-GameObject (generic)

This integration has not replaced any of the existing working code, but it will eventually.
I have managed to add MercSheets to the saving process, so it is available to completely convert into a MercSheet-only save system in the future