Roots now do damage when they expire with HP to spare
Roots can also deal damage each turn but that is disabled now -> need to added a "direct damage" options, since the "shield" component of the root (taking damage instead of the attached-to pawn) is taking the damage from the root (basically the root is hitting itself). Will be fixed soon!