Equip system added!
-IEquipable - interface, describes types of to access to the MagicItem data-class, to be used by the equipment system.

MercSheet got a new property!
-Gear, a property of MercSheet - basically holds all 3 item slots (and items therein).
  - Has methods to equip/unequip/swap items from those slots - also methods that gathers their benefits and allow easy access to item/benefit names, descriptions, sprites and specific access by slot type.

-MercGearDisplayer - enables the interaction with any given MercSheet's gear. This component utilizes the above mentions methods to swap/equip items.
Displays the mercs portrait, name, level - and has some spare room for data (min/max damage, max hp - so these stat changes would be instantly apparent and obvious).
And also holds 3 GearSlots to display that mercs items.

- GearSlots are the UI prefabs that display the currently equipped items from the selected merc's Gear. *This brought forth the issue of "more basic elements" - addressed in another document (see discord, progression)

GearSlots don't do much yet - they can only be "equipped to", but soon the ability to TRY simply unequipping an item (to be denied if that slot cannot remain empty)
Currently one can simply choose an item from the "Equip inventory" to fill an empty slot, or swap out an item in a slot, for one in the inventory.

- Equip Inventory, currently uses the bottom-left-most button the Tavern screen.
Comprised of 2 general windows:
 1) MercGearDisplayer - which takes a MercSheet, then displays that merc's Gear, loading their Gear.items into their respective slots.

  2) Inventories - also split to 2 Tabs:
 - Item inventory tab (shows all items)
 - Merc inventory tab (shows all mercs)

As of now, it simply displays your first merc's equipment, and the item inventory, plainly showing all items*** - each as a MagicItemDisplayer

- MagicItemDisplayer - UI element with 2 button functions, updated to function first as equipment (tries to equip on double-click)  - but still has the "sell item" mode which allows it to be sold. (this mode is only available via the old inventory window at the moment - it's not important where we sell items right now).
Double-clicking a MagicItemDisplayer will call the MercGearDisplayer to TryEquipItem for the currently selected merc (default selected = the first merc in your list).
The system knows which item goes to which slot, checks if there is an item present in that MercSheets gear[relevantSlot] - if so, it returns the removed item (so it could be added to the inventory again) - and removes the newly equipped item from the inventory.

-Static Class GeneralInputSettings - added to provide the waitTime for doubleclick (currently .5 seconds - should be less, and should be event based)



Next:
Tavern scene needs to learn to accept the new inventory.
- Include inventory in the "click nothing to close windows" function.
 -Static Class GeneralInputSettings - added to provide the waitTime for doubleclick needs to apply to all double clicks
GearSlots don't do much yet - they can only be "equipped to", but soon the ability to TRY simply unequipping an item (to be denied if that slot cannot remain empty)
- Research event based clickers for better double click.