CHANGES!
Mercs:
now use pawns as a scriptableobject.
Pawns hold MercSheets with the class and base stats of a character.
On a Pawns Init(), its mercSheetInPlayerData copies the baseStats from the prefab (also the MercClass) - then the Instantiated pawns sheet links to the mercSheetInPlayerData. [this may be silly since they are that way when instantiated]

Stat Blocks:
- BaseStatBlock, is Serialized on each Merc Prefab (loaded by Pawn.Init() )
- There is no "current/modifiedStatBlock" yet - but there will be.

SiteButtons:
sorting out the idea of "sending squads out" instead of just changing their status. Needs work.

TASKS:
- Add a "current/modifiedStatBlock" to Pawn. This will be the adjusted value of all stats after equipment!

MercSheetSOs may be useless