Many fixes, yes yes, good good
Esc to X - When an X button is enabled - it becomes the Current (meaning latest, most recently opened), if the Current is not null (meaning, an X button is still acitveAndEnabled when the new one gets enabled) - the new X button saves the old current to a private _previous, then becomes Current himself. When Esc is hit -> Current.onClick is invoked. If there is a _previous, that gets set to be the new Current.

- If Current is null, the normal EscMenu will open/close with Esc.

- This caused a bug where the Item Hover doesn't get a proper OnPointerExit -> in this case, hovers will be disabled with their items' OnDisable. SOLVED!

- SuqadBuilder2 has a new public static OnAnyCrewChanges action. CrewManager subs a refresh to it if the two are open simultaneously.

-


ClickNothingButtons were added.
- Hover Inventories have a dark-alpha panel behind them as to stop all other potential clicking.
- Tavern General ClickNothing was raised above the BG and now calls "DisableTier1Windows" which closes all the main windows


GearSlots - slot type circle was blocking the click, unticked Raycastable from image and text of the SlotType solved it.

- Temp, but sufficient ComparisonArrow pngs were added. (No need to colour or rotate)